-Refactoring -Split code into multiple files -Added tableaux generator
