+++ .gitignore
@@ -259,3 +259,5 @@ paket-files/
 # Python Tools for Visual Studio (PTVS)
 __pycache__/
 *.pyc
+/NDimYoung/Anim
+/NDimYoung/Anim1000
