+++ NDimYoung/Tableaux SQUARE DIM=9 CNT=10000.txt
@@ -1,0 +1,42 @@
+0 241
+1 250
+2 243
+3 253
+4 230
+5 255
+6 202
+7 248
+8 273
+9 228
+10 229
+11 224
+12 291
+13 255
+14 247
+15 249
+16 224
+17 229
+18 246
+19 255
+20 246
+21 236
+22 232
+23 206
+24 209
+25 219
+26 235
+27 251
+28 238
+29 240
+30 233
+31 233
+32 239
+33 270
+34 235
+35 242
+36 229
+37 218
+38 228
+39 239
+40 216
+41 234
