+++ NDimYoung/Tableaux TRIANGULAR DIM=10 CNT=10000.txt
@@ -1,0 +1,768 @@
+0 9
+1 10
+2 16
+3 17
+4 12
+5 16
+6 12
+7 17
+8 14
+9 19
+10 12
+11 14
+12 17
+13 21
+14 12
+15 14
+16 11
+17 16
+18 17
+19 10
+20 16
+21 12
+22 12
+23 10
+24 16
+25 18
+26 10
+27 20
+28 9
+29 16
+30 16
+31 15
+32 15
+33 14
+34 9
+35 13
+36 15
+37 18
+38 15
+39 14
+40 14
+41 13
+42 13
+43 19
+44 17
+45 13
+46 15
+47 18
+48 18
+49 15
+50 14
+51 8
+52 9
+53 19
+54 18
+55 12
+56 11
+57 18
+58 15
+59 10
+60 18
+61 9
+62 12
+63 9
+64 15
+65 16
+66 17
+67 8
+68 9
+69 16
+70 13
+71 16
+72 13
+73 14
+74 16
+75 14
+76 19
+77 11
+78 16
+79 11
+80 14
+81 11
+82 10
+83 15
+84 17
+85 10
+86 8
+87 10
+88 16
+89 12
+90 15
+91 21
+92 12
+93 16
+94 14
+95 8
+96 8
+97 14
+98 12
+99 13
+100 15
+101 13
+102 14
+103 19
+104 16
+105 15
+106 15
+107 20
+108 18
+109 12
+110 15
+111 13
+112 19
+113 16
+114 22
+115 16
+116 10
+117 19
+118 8
+119 18
+120 14
+121 16
+122 13
+123 17
+124 12
+125 13
+126 14
+127 16
+128 18
+129 18
+130 14
+131 5
+132 10
+133 11
+134 16
+135 17
+136 10
+137 9
+138 13
+139 16
+140 10
+141 16
+142 13
+143 12
+144 13
+145 13
+146 16
+147 8
+148 14
+149 20
+150 15
+151 11
+152 9
+153 11
+154 15
+155 14
+156 9
+157 7
+158 16
+159 14
+160 13
+161 12
+162 15
+163 13
+164 13
+165 11
+166 6
+167 11
+168 11
+169 8
+170 24
+171 12
+172 18
+173 10
+174 11
+175 10
+176 16
+177 16
+178 10
+179 17
+180 14
+181 14
+182 12
+183 14
+184 12
+185 16
+186 13
+187 17
+188 15
+189 18
+190 13
+191 13
+192 9
+193 15
+194 7
+195 11
+196 12
+197 9
+198 11
+199 17
+200 12
+201 10
+202 19
+203 11
+204 11
+205 14
+206 12
+207 12
+208 11
+209 11
+210 28
+211 13
+212 16
+213 10
+214 12
+215 15
+216 6
+217 16
+218 13
+219 13
+220 19
+221 17
+222 13
+223 11
+224 15
+225 13
+226 24
+227 19
+228 16
+229 17
+230 10
+231 12
+232 17
+233 10
+234 14
+235 8
+236 10
+237 13
+238 16
+239 14
+240 12
+241 18
+242 15
+243 11
+244 22
+245 16
+246 16
+247 22
+248 20
+249 14
+250 12
+251 18
+252 12
+253 10
+254 18
+255 9
+256 20
+257 15
+258 19
+259 11
+260 6
+261 15
+262 17
+263 13
+264 10
+265 11
+266 18
+267 8
+268 10
+269 16
+270 15
+271 9
+272 10
+273 11
+274 9
+275 12
+276 15
+277 13
+278 11
+279 18
+280 12
+281 10
+282 16
+283 18
+284 14
+285 16
+286 9
+287 12
+288 16
+289 18
+290 14
+291 14
+292 10
+293 18
+294 10
+295 10
+296 12
+297 13
+298 16
+299 17
+300 11
+301 14
+302 16
+303 12
+304 13
+305 14
+306 18
+307 10
+308 14
+309 8
+310 10
+311 19
+312 18
+313 13
+314 17
+315 14
+316 20
+317 12
+318 18
+319 17
+320 13
+321 10
+322 13
+323 14
+324 20
+325 10
+326 14
+327 14
+328 11
+329 17
+330 12
+331 14
+332 16
+333 11
+334 18
+335 15
+336 14
+337 20
+338 10
+339 13
+340 15
+341 12
+342 13
+343 17
+344 16
+345 14
+346 18
+347 21
+348 15
+349 11
+350 17
+351 14
+352 13
+353 7
+354 12
+355 14
+356 9
+357 14
+358 12
+359 11
+360 17
+361 14
+362 13
+363 17
+364 15
+365 13
+366 15
+367 13
+368 14
+369 11
+370 11
+371 16
+372 14
+373 12
+374 16
+375 8
+376 13
+377 13
+378 13
+379 6
+380 10
+381 13
+382 10
+383 20
+384 7
+385 14
+386 10
+387 4
+388 16
+389 19
+390 18
+391 13
+392 14
+393 16
+394 7
+395 17
+396 10
+397 19
+398 11
+399 17
+400 15
+401 13
+402 17
+403 11
+404 12
+405 10
+406 13
+407 21
+408 14
+409 8
+410 18
+411 11
+412 22
+413 10
+414 8
+415 12
+416 16
+417 11
+418 11
+419 12
+420 9
+421 18
+422 14
+423 16
+424 13
+425 16
+426 16
+427 10
+428 10
+429 12
+430 13
+431 15
+432 12
+433 10
+434 12
+435 11
+436 18
+437 19
+438 8
+439 13
+440 14
+441 26
+442 9
+443 14
+444 12
+445 12
+446 12
+447 14
+448 15
+449 12
+450 13
+451 11
+452 11
+453 12
+454 13
+455 20
+456 10
+457 11
+458 11
+459 12
+460 13
+461 12
+462 18
+463 16
+464 11
+465 16
+466 19
+467 18
+468 17
+469 14
+470 9
+471 14
+472 10
+473 11
+474 13
+475 16
+476 13
+477 14
+478 8
+479 10
+480 13
+481 15
+482 9
+483 14
+484 15
+485 19
+486 12
+487 14
+488 17
+489 10
+490 12
+491 13
+492 13
+493 11
+494 18
+495 16
+496 14
+497 14
+498 12
+499 14
+500 10
+501 16
+502 15
+503 13
+504 18
+505 14
+506 20
+507 18
+508 15
+509 13
+510 7
+511 20
+512 17
+513 14
+514 8
+515 19
+516 12
+517 15
+518 16
+519 11
+520 13
+521 11
+522 13
+523 17
+524 11
+525 7
+526 18
+527 16
+528 18
+529 12
+530 10
+531 14
+532 11
+533 8
+534 14
+535 12
+536 8
+537 17
+538 10
+539 9
+540 12
+541 16
+542 18
+543 14
+544 12
+545 13
+546 16
+547 13
+548 12
+549 9
+550 8
+551 16
+552 11
+553 14
+554 11
+555 12
+556 11
+557 12
+558 14
+559 12
+560 11
+561 8
+562 17
+563 13
+564 12
+565 16
+566 12
+567 14
+568 12
+569 8
+570 12
+571 12
+572 9
+573 16
+574 8
+575 7
+576 11
+577 17
+578 6
+579 13
+580 15
+581 7
+582 16
+583 6
+584 21
+585 11
+586 14
+587 13
+588 10
+589 9
+590 11
+591 11
+592 15
+593 8
+594 9
+595 9
+596 11
+597 11
+598 12
+599 17
+600 10
+601 11
+602 15
+603 16
+604 9
+605 14
+606 11
+607 11
+608 16
+609 12
+610 13
+611 12
+612 16
+613 12
+614 11
+615 6
+616 11
+617 12
+618 14
+619 4
+620 16
+621 10
+622 12
+623 11
+624 12
+625 9
+626 18
+627 9
+628 6
+629 15
+630 7
+631 10
+632 11
+633 17
+634 12
+635 12
+636 16
+637 12
+638 11
+639 14
+640 10
+641 11
+642 13
+643 14
+644 14
+645 14
+646 7
+647 6
+648 11
+649 14
+650 13
+651 8
+652 13
+653 10
+654 14
+655 14
+656 11
+657 10
+658 13
+659 8
+660 10
+661 12
+662 19
+663 10
+664 16
+665 14
+666 14
+667 12
+668 9
+669 14
+670 12
+671 11
+672 13
+673 13
+674 9
+675 16
+676 9
+677 13
+678 16
+679 9
+680 9
+681 12
+682 12
+683 15
+684 11
+685 14
+686 8
+687 11
+688 8
+689 8
+690 10
+691 16
+692 11
+693 11
+694 6
+695 6
+696 20
+697 7
+698 14
+699 10
+700 13
+701 8
+702 10
+703 12
+704 7
+705 15
+706 17
+707 9
+708 16
+709 10
+710 5
+711 10
+712 13
+713 13
+714 10
+715 8
+716 7
+717 8
+718 11
+719 9
+720 8
+721 10
+722 8
+723 15
+724 11
+725 8
+726 11
+727 9
+728 12
+729 9
+730 17
+731 9
+732 17
+733 9
+734 21
+735 9
+736 18
+737 12
+738 13
+739 12
+740 8
+741 13
+742 7
+743 18
+744 3
+745 10
+746 13
+747 16
+748 13
+749 18
+750 13
+751 9
+752 11
+753 16
+754 4
+755 6
+756 10
+757 7
+758 7
+759 7
+760 9
+761 6
+762 12
+763 8
+764 9
+765 3
+766 5
+767 7
